Implement word extrusion
Input: One word, one integer number (-100 <= N <100), L or R from the user. L or R can be entered in both upper and lower case letters.
If the given word is L, it pushes the given number of numbers to the left, and if it is R, it pushes it to the right.
Words that are pushed out are filled to the opposite side.
Input and output examples
Odd lines are input, even lines are output.
> apple 3 L
leapp

> banana 6 R
banana

> carrot -1 r
arrotc

> cat -4 R
atc